requirements.txt
confusable-homoglyphs==3.0.0
dj-database-url==0.5.0
Django==1.11.7
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-crispy-forms-registration==0.1.3
django-avatar==4.1.0
django-registration==2.4.1
django-registration-redux==2.2
django-bootstrap3-datetimepicker-2==2.5.0
django-bootstrap-form==3.4
django-javascript-settings==1.1.2
